requirements.txt
azure-ai-projects
azure-identity
python-dotenv
openai

dotenv
matplotlib
pillow
azure-ai-vision-imageanalysis==1.0.0